Remove spaces in GitLab integration details
In the GitLab integration details, Prettier was making some changes that introduced extra spaces and made Hugo render the markdown in the "Off by default" section as a separate paragraph. This is me attempting to remove the spaces manually and see if that fixes the way the markdown rendering creates the HTML, and therefore fixes the formatting issue.